support multiple named mods / dependencies in single manifest
Support cases where a single mod download has a manifest that bundles
multiple mods, rather than (the usual) just one. Example: Arhpy's
D-Sides.

Assumes that any additional named mods in the manifest have a "DLL"
item (as in Arphy), or at least that the "main" mod that bundles the
others (e.g. the one the user intended to download) is listed first
in the manifest. Otherwise, it is ambiguous from a downloaded mod
package what mod the download is supposed to represent, so we don't
know where to install it.